move the updateRects to its own file and simplify the code
